requirements.txt
fastapi
uvicorn[standard]

# база данных
sqlalchemy
databases
pydantic
databases

# .env
python-dotenv

passlib
pillow
python-multipart

dicom2jpg
onnxruntime
opencv-python
numpy
ultralytics

# webrtc
websockets
starlette